Refactored the mock service registry to be a common memory service registry
